+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/type/EpisodeViewSource.kt
@@ -10,6 +10,7 @@ enum class EpisodeViewSource(val value: String) {
     UP_NEXT("up_next"),
     SHARE("share"),
     NOTIFICATION("notification"),
+    NOTIFICATION_BOOKMARK("notification_bookmark"),
     SEARCH("search"),
     SEARCH_HISTORY("search_history"),
     UNKNOWN("unknown");